requirements.txt
flask>=3.1.1
flask_sqlalchemy>=3.1.1
sqlalchemy>=2.0.41
requests>=2.32.4
beautifulsoup4>=4.13.4
trafilatura>=2.0.0
openai>=1.86.0
psycopg2-binary>=2.9.10
python-telegram-bot>=22.1
email-validator>=2.2.0
gunicorn>=23.0.0
python-dotenv>=1.0.0
werkzeug>=3.1.3
zoneinfo; python_version<'3.9'
urllib3<2.0.0
